+++ lua/kustomize/init.lua
@@ -3,6 +3,7 @@ local M = {}
 local build = require("kustomize.build")
 local kinds = require("kustomize.kinds")
 local open = require("kustomize.open")
+local resources = require("kustomize.resources")
 
 M.build = function()
   build.build()
@@ -16,6 +17,10 @@ M.open = function()
   open.open()
 end
 
+M.resources = function()
+  resources.resources()
+end
+
 M.setup = function()
   vim.keymap.set("n", "<leader>kb", function()
     M.build()
@@ -26,6 +31,9 @@ M.setup = function()
   vim.keymap.set("v", "<leader>ko", function()
     M.open()
   end, { desc = "Open" })
+  vim.keymap.set("n", "<leader>kr", function()
+    M.resources()
+  end, { desc = "Print resources in folder" })
 end
 
 return M