+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-11-21
+
+### Added
+
+- "resources" use case to quickly add all files/folders within a folder containing a kustomization.ya(m)l to the list of resources
+
+### Fixed
+
+- vim commands
+
+### Changed
+
+- refactor check for `kustomization.y(a)ml` to a function in utils
+
 ## [0.2.1] - 2022-11-20
 
 ### Added